aider: Update available times when booking date changes.
# Aider chat conversation:

ASSISTANT: You're correct. Streamlit doesn't support callbacks like some other frameworks, but we can simulate this behavior by using the `st.session_state` object to store the state of the application and trigger updates when necessary.

We can create a function that will be called every time the page is re-rendered. This function will check if the `booking_date` has changed and, if so, update the `available_times`.

Here is the change:

```python
main.py
<<<<<<< HEAD
        booking_date = st.date_input('Booking Date', min_value=min_date, max_value=max_date, format="DD.MM.YYYY")
        if 'booking_date' not in st.session_state or st.session_state['booking_date'] != booking_date:
            st.session_state['booking_date'] = booking_date
            available_times = get_available_times(convert_date_format(booking_date))
        else:
            available_times = get_available_times(convert_date_format(st.session_state['booking_date']))
=======
        booking_date = st.date_input('Booking Date', min_value=min_date, max_value=max_date, format="DD.MM.YYYY")
        st.session_state['booking_date'] = booking_date
        if 'available_times' not in st.session_state:
            st.session_state['available_times'] = get_available_times(convert_date_format(booking_date))
        else:
            if st.session_state['booking_date'] != booking_date:
                st.session_state['available_times'] = get_available_times(convert_date_format(booking_date))
        available_times = st.session_state['available_times']
>>>>>>> updated
```

This change will ensure that the `available_times` are updated whenever the `booking_date` changes.